src/backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
geoalchemy2==0.15.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bleach==6.2.0
slowapi==0.1.9

# Object Storage
aioboto3==13.2.0

# Image Processing
Pillow==11.1.0

# Configuration
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
locust==2.32.2

# Email
fastapi-mail==1.4.1
